+++ CHANGELOG.rst
@@ -9,7 +9,7 @@ Version 2.0
 - Issue #14 is fixed
 - Parameter ``empty_lines_in_values`` is now activated by default and can be changed
 - Renamed ``sections_blocks`` to ``section_blocks`` for consistency
-- ``last_item`` for renamed to ``last_block`` for consistency
+- ``last_item`` renamed to ``last_block`` for consistency
 - Reworked some internal parts of the inheritance hierarchy
 - added ``remove`` to remove the current block
 - added ``next_block`` and ``previous_block`` for easier navigation in section